Split uwp platform export template into multiple files
